pyproject.toml
[project]
name = "indonesia-supreme-court-llm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "langchain-community>=0.3.10",
    "langchain-core>=0.3.23",
    "langchain-openai>=0.2.12",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.2.59",
    "litellm>=1.54.1",
    "pydantic>=2.10.3",
    "sqlmodel>=0.0.22",
    "tenacity>=9.0.0",
    "typer>=0.15.1",
]
